+++ R/add.R
@@ -188,8 +188,31 @@ addFeatures <- function(study, features, reset = FALSE) {
 #'
 #' @param models The models analyzed in the study. The input is a named list.
 #'   The names correspond to the names of the models. The elements correspond to
-#'   the descriptions of the models.
+#'   the descriptions of the models. Alternatively, instead of a single
+#'   character string, you can provide a list of metadata fields about each
+#'   model. The field "description" will be used to derive the tooltip displayed
+#'   in the app.
 #' @inheritParams shared-add
+#'
+#' @examples
+#'   study <- createStudy("example")
+#'   models <- list(
+#'     model_01 = "Name of first model",
+#'     model_02 = "Name of second model"
+#'   )
+#'   study <- addModels(study, models)
+#'
+#'   # Alternative: provide additional metadata about each model
+#'   models <- list(
+#'     model_01 = list(
+#'       description = "Name of first model",
+#'       data_type = "transcriptomics"
+#'     ),
+#'     model_02 = list(
+#'       description = "Name of second model",
+#'       data_type = "proteomics"
+#'     )
+#'   )
 #'
 #' @export
 addModels <- function(study, models, reset = FALSE) {